moved compile commands w Shreyas
